+++ index.js
@@ -49,7 +49,7 @@ module.exports = (context, options = {}) => {
     corejs: 2,
     helpers: true,
     regenerator: true,
-    version: '7.22.0', // 1
+    version: '7.23.0', // 1
   };
 
   let presets = [